requirements.txt
langgraph
langchain
langmem
langchain-openai
langchain_core
langchain-google-genai
pillow
colpali_engine
huggingface_hub[hf_transfer]
transformers==4.53.0
qdrant-client
fsspec
protobuf
openai
httpx
psycopg2-binary
asyncio
langgraph-checkpoint-postgres
google-genai
google-cloud-storage
fastapi[standard]
uvicorn
pydantic
rich